ci: fix merge
