+++ .fvm/flutter_sdk
@@ -1,0 +1,1 @@
+C:/Users/Alvaro Vasconcelos/fvm/versions/stable
