+++ modular.code-workspace
@@ -41,7 +41,7 @@
 		}
 	],
 	"settings": {
-		"dart.flutterSdkPath": "/Users/jacobmoura/fvm/versions/stable",
+		"dart.flutterSdkPath": ".fvm\flutter_sdk",
 		"lcov.branchCoverage": "off",
 		"lcov.path": "coverage/lcov.info"
 	}
